Add files via upload
A report on my attempts to solve this simple wall-following problem as effectively as
possible using a form of evolutionary intelligence Genetic Algorithm:
The simulated robot has five sensors which can detect the presence or absence of wall in the five
grid locations it faces, as shown. It can take one of four actions: do nothing, turn left or right (at
90 degrees) on the spot, or go forward one cell. The fitness of the robot is how many of the
highlighted cells next to the wall it has covered in a fixed number of sense- act cycles. Robots
are allowed 28 such cycles per trial, i.e., enough to follow the wall perfectly. The robot starts in
the position indicated, facing East.